Create empty react project and install lingui
