+++ lingui/.babelrc
@@ -1,0 +1,8 @@
+{
+    "plugins": [
+        "macros"
+    ],
+    "presets": [
+        "@babel/preset-react"
+    ]
+  }
