+++ lingui/.linguirc
@@ -1,0 +1,8 @@
+{
+   "locales": ["en", "cs"],
+   "catalogs": [{
+      "path": "src/locales/{locale}/messages",
+      "include": ["src"]
+   }],
+   "format": "minimal"
+}
