+++ lingui/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "lingui-demo",
+  "version": "0.0.1",
+  "description": "Demo of lingui.js library in React Project",
+  "main": "index.js",
+  "license": "MIT",
+  "private": false,
+  "scripts": {
+    "build": "webpack --mode production",
+    "start": "webpack serve --mode development --open",
+    "i18n:extract": "lingui extract",
+    "i18n:compile": "lingui compile"
+  },
+  "dependencies": {
+    "@lingui/react": "^3.13.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.17.9",
+    "@babel/preset-react": "^7.16.7",
+    "@lingui/cli": "^3.13.2",
+    "@lingui/macro": "^3.13.2",
+    "babel-loader": "^8.2.4",
+    "babel-plugin-macros": "^3.1.0",
+    "html-webpack-plugin": "^5.5.0",
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.8.1"
+  }
+}
