+++ lingui/src/App.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const App = () => (
+    <div>
+        <h1>React from Scratch</h1> 
+    </div>
+);
+
+export default App;
