+++ lingui/src/index.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import ReactDOMClient from "react-dom/client";
+
+import App from "./App";
+
+const root = ReactDOMClient.createRoot(document.querySelector("#root"));
+
+root.render(<App />);
