+++ lingui/src/locales/cs/messages.json
@@ -1,0 +1,1 @@
+{}
