+++ lingui/src/locales/en/messages.json
@@ -1,0 +1,1 @@
+{}
