--- package.json
@@ -1,26 +1,0 @@
-{
-  "name": "react-i18n-lingui-demo",
-  "version": "1.0.0",
-  "main": "index.tsx",
-  "license": "MIT",
-  "devDependencies": {
-    "@types/react": "^18.0.3",
-    "@types/react-dom": "^18.0.0",
-    "css-loader": "^6.7.1",
-    "html-webpack-plugin": "^5.5.0",
-    "mini-css-extract-plugin": "^2.6.0",
-    "ts-loader": "^9.2.8",
-    "typescript": "^4.6.3",
-    "webpack": "^5.72.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.8.1"
-  },
-  "scripts": {
-    "start": "webpack serve --port 3000",
-    "build": "NODE_ENV=production webpack"
-  },
-  "dependencies": {
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0"
-  }
-}
