--- public/index.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head lang="en">
-  <title>Lingui React</title>
-</html>
-<body>
-  <div id="app-root">App is loading...</div>
-</body>
