--- src/App.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const App = () => (
-    <div className='container'>
-        <h1>Hello World</h1>
-    </div>
-)
-
-export default App;
