--- src/index.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-import * as ReactDOMClient from 'react-dom/client';
-
-import App from './App';
-
-const root = ReactDOMClient.createRoot(document.getElementById('app-root')!);
-
-root.render(<App/>);
