--- tsconfig.json
@@ -1,18 +1,0 @@
-{
-    "compilerOptions": {
-      "esModuleInterop": true,
-      "jsx": "react",
-      "module": "esnext",
-      "moduleResolution": "node",
-      "lib": [
-        "dom",
-        "esnext",
-      ],
-      "strict": true,
-      "sourceMap": true,
-      "target": "esnext",
-    },
-    "exclude": [
-      "node_modules"
-    ]
-  }
